# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tmux-gateway",
    "description": "REST API for interacting with local tmux sessions",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Health check",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "get": {
        "tags": [],
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "List tmux sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to list sessions"
          }
        }
      },
      "post": {
        "tags": [],
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewSessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create session"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "NewSessionRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "NewSessionResponse": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "name",
          "windows",
          "created",
          "attached"
        ],
        "properties": {
          "attached": {
            "type": "boolean"
          },
          "created": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "windows": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      }
    }
  }
}